+++ CHANGELOG.md
@@ -14,10 +14,11 @@ All user visible changes to this project will be documented in this file. This p
 ### Changed
 
 - Refactor Audio Device Module to use [OpenAL] library for playout. ([#117])
-- Fire `onDeviceChange` callback whenever an output audio device is changed in system settings on desktop platforms. ([#119])
+- Fire `onDeviceChange` callback whenever an output audio device is changed in system settings on desktop platforms. ([#119], [#120])
 
 [#117]: https://github.com/instrumentisto/medea-flutter-webrtc/pull/117
 [#119]: https://github.com/instrumentisto/medea-flutter-webrtc/pull/119
+[#120]: https://github.com/instrumentisto/medea-flutter-webrtc/pull/120